Send log to nelo (#138)
Currently, lambda has saved its logs to local disk, in order to manage log
together, it is necessary to send these logs to nelo.

In lambda-dev/real env, we deploy all components to different hosts,
so it is necessary to create softlink or hardlink to point all logs to
same directory which neloagent collects the logs.

Due to neloagent doesn't support softlink, so we use hardlink instead.
On the other hand, neloagent will upgrade from c++ to java, so in future,
may simplify our neloagent configuration.

Resolved: paas/lambda#91