Enable proxy protocol on port 8443 (#208)
As the ncloud LB doesn't support DSR, we need enable proxy protocol
in nginx to get the user's real ip, and we also need the nginx be
accessed directly, so we use port 8443 for the proxy protocol as
this port is exposed in the nginx Dockerfile but not used right now.
Add https suport for 8443 port.